fix: typos and filenames
