+++ website/src/pages/StatusPage/components/PreviousExecutions/ExecutionTable.tsx
@@ -17,6 +17,8 @@ limitations under the License.
 import {
   ColumnDef,
   ColumnFiltersState,
+  SortingState,
+  VisibilityState,
   flexRender,
   getCoreRowModel,
   getFacetedRowModel,
@@ -24,12 +26,10 @@ import {
   getFilteredRowModel,
   getPaginationRowModel,
   getSortedRowModel,
-  SortingState,
   useReactTable,
-  VisibilityState,
 } from "@tanstack/react-table";
+import * as React from "react";
 
-import { DataTablePagination } from "@/components/ui/data-table-pagination";
 import {
   Table,
   TableBody,
@@ -38,25 +38,26 @@ import {
   TableHeader,
   TableRow,
 } from "@/components/ui/table";
-import React from "react";
-import { DataTableToolbar } from "../PreviousExecutions/PreviousExecutionsTableToolbar";
 
-interface DataTableProps<ExecutionQueueType, TValue> {
-  columns: ColumnDef<ExecutionQueueType, TValue>[];
-  data: ExecutionQueueType[];
+import { DataTablePagination } from "@/components/ui/data-table-pagination";
+import { DataTableToolbar } from "./ExecutionTableToolbar";
+
+interface DataTableProps<TData, TValue> {
+  columns: ColumnDef<TData, TValue>[];
+  data: TData[];
 }
 
-export function ExexutionQueueTable<ExecutionQueueType, TValue>({
+export function PreviousExecutionQueueTable<TData, TValue>({
   columns,
   data,
-}: DataTableProps<ExecutionQueueType, TValue>) {
-  const [rowSelection, setRowSelection] = React.useState({})
+}: DataTableProps<TData, TValue>) {
+  const [rowSelection, setRowSelection] = React.useState({});
   const [columnVisibility, setColumnVisibility] =
-    React.useState<VisibilityState>({})
+    React.useState<VisibilityState>({});
   const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>(
     []
-  )
-  const [sorting, setSorting] = React.useState<SortingState>([])
+  );
+  const [sorting, setSorting] = React.useState<SortingState>([]);
 
   const table = useReactTable({
     data,
@@ -90,7 +91,10 @@ export function ExexutionQueueTable<ExecutionQueueType, TValue>({
               <TableRow key={headerGroup.id} className="border-border">
                 {headerGroup.headers.map((header) => {
                   return (
-                    <TableHead key={header.id} className="text-center font-semibold text-foreground">
+                    <TableHead
+                      key={header.id}
+                      className="text-center font-semibold text-foreground"
+                    >
                       {header.isPlaceholder
                         ? null
                         : flexRender(
